backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
semantic-kernel>=0.4.0
python-dotenv>=1.0.0
pyodbc>=4.0.39
pandas>=2.1.1
numpy>=1.24.0
nest-asyncio>=1.5.8
azure-identity>=1.14.1
azure-ai-projects==1.0.0b10
streamlit>=1.28.0
requests>=2.31.0
plotly>=5.13.0
azure-storage-blob>=12.19.0
torch>=2.0.0
scikit-learn>=1.3.0
scipy>=1.11.0
python-docx>=0.8.11
markdown>=3.5.1
aiohttp>=3.9.0
pydantic>=2.0.0
